+++ components/com_kinoarhiv/layouts/editors/editor_trumbowyg.php
@@ -1,0 +1,101 @@
+<?php
+/**
+ * @package     Kinoarhiv.Site
+ * @subpackage  com_kinoarhiv
+ *
+ * @copyright   Copyright (C) 2010 Libra.ms. All rights reserved.
+ * @license     GNU General Public License version 2 or later
+ * @url            http://киноархив.com/
+ */
+
+defined('_JEXEC') or die;
+
+JHtml::_('script', 'components/com_kinoarhiv/assets/editors/trumbowyg/trumbowyg.min.js');
+JHtml::_('stylesheet', 'components/com_kinoarhiv/assets/editors/trumbowyg/ui/default.css');
+KAComponentHelper::getScriptLanguage('', true, 'editors/trumbowyg');
+
+$params = $displayData->params;
+$form   = $displayData->form;
+?>
+<script type="text/javascript">
+	jQuery(document).ready(function($){
+		var editor = $('#form_review').trumbowyg({
+			lang: '<?php echo substr(JFactory::getLanguage()->getTag(), 0, 2); ?>',
+			removeformatPasted: true,
+			btns: [ 'formatting', '|', 'btnGrp-design', '|', 'btnGrp-justify', '|', 'btnGrp-lists', '|', 'viewHTML' ]
+		});
+
+		// Insert username into editor
+		$('.cmd-insert-username').click(function(){
+			var username = $(this).text();
+
+			editor.trumbowyg('html', '<strong>' + username + '</strong><br />');
+		});
+
+		// Insert cite into editor
+		$('.cmd-insert-quote').click(function(e){
+			e.preventDefault();
+
+			var review = $(this).closest('.review-row');
+			var quoted_text = review.find('.review').html(),
+				quoted_link = review.find('.review-row-title a.permalink').attr('href'),
+				username = review.find('.review-row-title span.username').text();
+
+			editor.trumbowyg(
+				'html',
+				'<a href="' + quoted_link + '"><strong>' + username + '</strong><?php echo JText::_('COM_KA_REVIEWS_QUOTEWROTE'); ?>:</a>'
+					+ '<br /><blockquote cite="' + quoted_link + '">' + quoted_text + '</blockquote><br />'
+			);
+		});
+
+		$('.cmd-reset').click(function(){
+			editor.trumbowyg('empty');
+
+			return true;
+		});
+
+		$('form.editor').submit(function(e){
+			var editor_text = editor.trumbowyg('html'),
+				min_length = <?php echo (int) $params->get('reviews_length_min'); ?>,
+				max_length = <?php echo (int) $params->get('reviews_length_max'); ?>,
+				submit = $('input[type="submit"]', this);
+
+			submit.attr('disabled', true);
+
+			if (editor_text.length < min_length || editor_text.length > max_length) {
+				showMsg(
+					$('.cmd-reset', this),
+					'<?php echo JText::sprintf(
+						JText::_('COM_KA_EDITOR_EMPTY'),
+						(int) $params->get('reviews_length_min'),
+						(int) $params->get('reviews_length_max')
+					); ?>'
+				);
+
+				window.setTimeout(function(){
+					submit.removeAttr('disabled');
+				}, 5000);
+
+				return false;
+			} else {
+				submit.removeAttr('disabled');
+			}
+
+			return true;
+		});
+	});
+</script>
+<div style="clear: both;">&nbsp;</div>
+<form action="<?php echo htmlspecialchars(JURI::getInstance()->toString()); ?>" method="post" id="review-form" class="uk-form editor">
+	<p><?php echo $form->getInput('review'); ?></p>
+	<div class="select-type"><?php echo $form->getLabel('type'); ?><?php echo $form->getInput('type'); ?></div>
+	<div class="clear"></div>
+	<?php
+	echo $form->getInput('captcha');
+	echo JHtml::_('form.token'); ?>
+	<input type="hidden" name="task" id="task" value="<?php echo $displayData->task; ?>"/>
+	<input type="hidden" name="id" value="<?php echo $displayData->id; ?>"/>
+	<br/>
+	<input type="submit" class="btn btn-primary uk-button uk-button-primary" value="<?php echo JText::_('JSUBMIT'); ?>"/>
+	<input type="reset" class="btn btn-default uk-button cmd-reset" value="<?php echo JText::_('JSEARCH_FILTER_CLEAR'); ?>"/>
+</form>